requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
